requirements.txt
pandas
praw
python-dotenv
nltk
textblob
matplotlib
seaborn
wordcloud